% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{account_info}
\alias{account_info}
\title{Get information about your account}
\usage{
account_info()
}
\value{
A list containing information about account
}
\description{
Get information about your account
}
\examples{
\dontrun{
ainfo <- account_info()
names(ainfo)
ainfo$Latest_Submission
}
}
