% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{binded_user_data}
\alias{binded_user_data}
\title{Get the performance of the user over time aggregated}
\usage{
binded_user_data(username, dates = NULL, merge = FALSE)
}
\arguments{
\item{username}{A vector of one or more usernames}

\item{dates}{A vector of dates to consider}

\item{merge}{Whether to merge multiple usernames into one}
}
\description{
Get the performance of the user over time aggregated
}
