% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{current_round}
\alias{current_round}
\title{Get current round and it's closing time}
\usage{
current_round()
}
\value{
Returns the current round number and it's closing times
}
\description{
Get current round and it's closing time
}
\examples{
\dontrun{
current_round()
}
}
