% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseapi.R
\name{daily_submission_performances}
\alias{daily_submission_performances}
\title{Fetch Daily Submission Performance of any user}
\usage{
daily_submission_performances(username, tournament = 8)
}
\arguments{
\item{username}{User Name}

\item{tournament}{Tournament ID, 8 for Main, 11 for Signal}
}
\value{
list of round model performance entries
}
\description{
Fetch Daily Submission Performance of any user
}
\examples{
\dontrun{
daily_submission_performances(username = "bayo")
}
}
