% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{download_data}
\alias{download_data}
\title{Function to download the Numerai Tournament data}
\usage{
download_data(
  location = tempdir(),
  legacy = TRUE,
  load = TRUE,
  live_only = FALSE
)
}
\arguments{
\item{location}{The directory path in which to store the data}

\item{legacy}{logical; new dataset.  If \code{legacy = TRUE}, legacy 310 feature .zip dataset downloaded}

\item{load}{logical; to unzip / read.csv or read_parquet on data.}

\item{live_only}{logical; download only the live .parquet dataset.}
}
\value{
A list containing the training and tournament data objects
}
\description{
Function to download the Numerai Tournament data
}
\examples{
\dontrun{
## Directory where data files and prediction files to be saved
## Put custom directory path or use the current working directory
data_dir <- tempdir()

## Download legacy dataset for current competition
data <- download_data(data_dir)
data_train <- data$data_train
data_tournament <- data$data_tournament

## Download super massive dataset for current competition
download(dat_dir, legacy = FALSE)
}
}
