% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numerapi.R
\name{download_dataset}
\alias{download_dataset}
\title{Download specified file for the given round.}
\usage{
download_dataset(
  filename = "v2/numerai_live_data.csv",
  dest_path = NA,
  round_num = NA
)
}
\arguments{
\item{filename}{File to be downloaded, defaults to live data}

\item{dest_path}{File path where the file should be stored, Defaults to current directory with filename parameter}

\item{round_num}{Tournament round you are interested in, defaults to the current round.}
}
\value{
list of filenames
}
\description{
Download specified file for the given round.
}
\examples{
\dontrun{
download_dataset(round_num=328)
}
}
