% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signalsapi.R
\name{download_validation_data}
\alias{download_validation_data}
\title{Download CSV file with historical targets and ticker universe}
\usage{
download_validation_data(file_path = "signals_historical_targets.csv")
}
\arguments{
\item{file_path}{CSV file with predictions that will get uploaded}
}
\value{
List of currently accepted tickers
}
\description{
Download CSV file with historical targets and ticker universe
}
\examples{
\dontrun{
download_validation_data(file_path = "signals_historical_targets.csv")
}
}
