% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseapi.R
\name{get_account}
\alias{get_account}
\title{Get all information about your account}
\usage{
get_account()
}
\value{
User information
}
\description{
Get all information about your account
}
\examples{
\dontrun{
get_account()
}
}
