% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{get_api_key}
\alias{get_api_key}
\title{Gets the Numerai API key}
\usage{
get_api_key()
}
\value{
Your Numerai API key, if set
}
\description{
Gets the Numerai API key
}
\examples{
\dontrun{
get_api_key()
}
}
