% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseapi.R
\name{get_competitions}
\alias{get_competitions}
\title{Retrieves information about all rounds}
\usage{
get_competitions(tournament = 8)
}
\arguments{
\item{tournament}{Tournament ID, 8 for Main, 11 for Signal}
}
\value{
Rounds Information
}
\description{
Retrieves information about all rounds
}
\examples{
\dontrun{
get_competitions()
}
}
