% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseapi.R
\name{get_current_round}
\alias{get_current_round}
\title{Get number of the current active round.}
\usage{
get_current_round()
}
\value{
Number of the current active round
}
\description{
Get number of the current active round.
}
\examples{
\dontrun{
get_current_round()
}
}
