% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseapi.R
\name{get_leaderboard}
\alias{get_leaderboard}
\title{Get the current leaderboard}
\usage{
get_leaderboard(tournament = 8)
}
\arguments{
\item{tournament}{Tournament ID, 8 for Main, 11 for Signal}
}
\value{
Leaderboard entries
}
\description{
Get the current leaderboard
}
\examples{
\dontrun{
get_leaderboard()
}
}
