% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseapi.R
\name{get_models}
\alias{get_models}
\title{Get mapping of account model names to model ids for convenience}
\usage{
get_models(tournament = 8)
}
\arguments{
\item{tournament}{Tournament ID, 8 for Main, 11 for Signal}
}
\value{
modelname -> model_id list
}
\description{
Get mapping of account model names to model ids for convenience
}
\examples{
\dontrun{
get_models()
}
}
