% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{get_password}
\alias{get_password}
\title{Gets the Numerai Password}
\usage{
get_password()
}
\value{
Your Numerai Password, if set
}
\description{
Gets the Numerai Password
}
\examples{
\dontrun{
get_password()
}
}
