% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{get_public_id}
\alias{get_public_id}
\title{Gets the Numerai Public ID}
\usage{
get_public_id()
}
\value{
Your Numerai Public ID, if set
}
\description{
Gets the Numerai Public ID
}
\examples{
\dontrun{
get_public_id()
}
}
