% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numerapi.R
\name{list_datasets}
\alias{list_datasets}
\title{List of available data files}
\usage{
list_datasets(round_num = NA)
}
\arguments{
\item{round_num}{Tournament round you are interested in, defaults to the current round.}
}
\value{
list of filenames
}
\description{
List of available data files
}
\examples{
\dontrun{
list_datasets(round_num=328)
}
}
