% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{release_nmr}
\alias{release_nmr}
\title{Release NMR}
\usage{
release_nmr(value, mfa_code = "", password = "")
}
\arguments{
\item{value}{The amount of NMR to release}

\item{mfa_code}{The mfa code}

\item{password}{Your password}
}
\value{
The transaction hash for release request
}
\description{
Release NMR
}
\examples{
\dontrun{
release_tx_hash <- release_nmr(value = 1)
}
}
