% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{run_query}
\alias{run_query}
\title{Function to run a raw GraphQL query on the API interface}
\usage{
run_query(query, id = get_public_id(), key = get_api_key())
}
\arguments{
\item{query}{The graphQL query to run on the API as a string in single quotes}

\item{id}{The public id of the Numerai application}

\item{key}{The Numerai API key}
}
\value{
The parsed json content returned from the request
}
\description{
Function to run a raw GraphQL query on the API interface
}
\examples{
\dontrun{
## Run Custom GraphQL code from R
custom_query <- "query queryname {
rounds (number:82) {
closeTime
}
}"
run_query(query=custom_query)$data
}
}
