% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseapi.R
\name{set_api_key}
\alias{set_api_key}
\title{Sets the Numerai API key}
\usage{
set_api_key(key)
}
\arguments{
\item{key}{The Numerai API key}
}
\value{
A boolean TRUE if the key was successfully set
}
\description{
Sets the Numerai API key
}
\examples{
\dontrun{
set_api_key("abcdefghijklmnop")
}
}
