% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseapi.R
\name{set_link}
\alias{set_link}
\title{Set link field for a model id}
\usage{
set_link(model_id, link, link_text)
}
\arguments{
\item{model_id}{Target model UUID}

\item{link}{URL}

\item{link_text}{URL Text}
}
\value{
If the link was changed successfully
}
\description{
Set link field for a model id
}
\examples{
\dontrun{
set_link(model_id = get_models()[["bayo"]], link = "https://www.google.com",link_text = "Google")
}
}
