% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{set_password}
\alias{set_password}
\title{Sets the Numerai Password}
\usage{
set_password(pass)
}
\arguments{
\item{pass}{The Numerai Password}
}
\value{
A boolean TRUE if the password was successfully set
}
\description{
Sets the Numerai Password
}
\examples{
\dontrun{
set_password("abcdefghijklmnop")
}
}
