% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{set_public_id}
\alias{set_public_id}
\title{Sets the Numerai Public ID}
\usage{
set_public_id(id)
}
\arguments{
\item{id}{The Numerai Public ID}
}
\value{
A boolean TRUE if the ID was successfully set
}
\description{
Sets the Numerai Public ID
}
\examples{
\dontrun{
set_public_id("abcdefghijklmnop")
}
}
