% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseapi.R
\name{set_stake_type}
\alias{set_stake_type}
\title{Change stake type by model}
\usage{
set_stake_type(
  model_id,
  corr_multiplier = 0,
  tc_multiplier = 0,
  take_profit = FALSE,
  tournament = 8
)
}
\arguments{
\item{model_id}{Target model UUID}

\item{corr_multiplier}{Multiplier of correlation for returns (Integer)}

\item{tc_multiplier}{Multiplier of TC for returns (Numeric)}

\item{take_profit}{TRUE/FALSE Determines whether payouts are returned to user wallet or automatically staked to next round.}

\item{tournament}{Tournament ID, 8 for Main, 11 for Signal}
}
\value{
Confirmation that payout selection has been updated
}
\description{
Change stake type by model
}
\examples{
\dontrun{
set_stake_type(model_id = get_models()[["bayo"]],corr_multiplier=1,tc_multiplier=3,tournament=8)
}
}
