% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseapi.R
\name{set_submission_webhook}
\alias{set_submission_webhook}
\title{Set a model's submission webhook used in Numerai Compute}
\usage{
set_submission_webhook(model_id, webhook)
}
\arguments{
\item{model_id}{Target model UUID}

\item{webhook}{The compute webhook to trigger this model}
}
\value{
Confirmation that your webhook has been set
}
\description{
Set a model's submission webhook used in Numerai Compute
}
\examples{
\dontrun{
set_submission_webhook(model_id = get_models()[["bayo"]], webhook = "..")
}
}
