% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseapi.R
\name{stake_change}
\alias{stake_change}
\title{Change stake by `value` NMR}
\usage{
stake_change(nmr, action = "decrease", model_id, tournament = 8)
}
\arguments{
\item{nmr}{amount of NMR you want to increase/decrease}

\item{action}{`increase` or `decrease`}

\item{model_id}{Target model UUID}

\item{tournament}{Tournament ID, 8 for Main, 11 for Signal}
}
\value{
Stake change information
}
\description{
Change stake by `value` NMR
}
\examples{
\dontrun{
stake_change(nmr=10,action="decrease",model_id = get_models()[["bayo"]],tournament=8)
}
}
