% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{stake_nmr}
\alias{stake_nmr}
\title{Stake NMR}
\usage{
stake_nmr(value, model_id = NULL, mfa_code = "", password = "")
}
\arguments{
\item{value}{The amount of NMR to stake}

\item{model_id}{The id of the model with which to stake}

\item{mfa_code}{The mfa code}

\item{password}{Your password}
}
\value{
The transaction hash for stake made
}
\description{
Stake NMR
}
\examples{
\dontrun{
stake_tx_hash <- stake_nmr(value = 1)
}
}
