% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{stake_nmr_multi}
\alias{stake_nmr_multi}
\title{Stake NMR on the current round and multiple tournaments}
\usage{
stake_nmr_multi(tournaments, values, confidence_vals, mfa_code = "",
  password = "")
}
\arguments{
\item{tournaments}{The vector of names of the tournaments}

\item{values}{The vector of the amounts of NMR to stake}

\item{confidence_vals}{The vector of the confidence values to use}

\item{mfa_code}{The mfa code}

\item{password}{Your password}
}
\value{
The transaction hashes for stakes made
}
\description{
Stake NMR on the current round and multiple tournaments
}
\examples{
\dontrun{
stake_tx_hashes <- stake_nmr_multi(tournaments=c("Bernie","Frank"),values = c(1,1), confidence_vals = c(".25",".5"))
}
}
