% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{status_submission_by_id}
\alias{status_submission_by_id}
\title{Get information about a submission from a submission id}
\usage{
status_submission_by_id(sub_id)
}
\arguments{
\item{sub_id}{The id of the submission}
}
\value{
A list containing information about the given submission id
}
\description{
Get information about a submission from a submission id
}
\examples{
\dontrun{
status_submission_by_id(submission_id)
}
}
