% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseapi.R
\name{submission_status}
\alias{submission_status}
\title{Submission status of the last submission associated with the account}
\usage{
submission_status(model_id, tournament = 8)
}
\arguments{
\item{model_id}{Target model UUID}

\item{tournament}{Tournament ID, 8 for Main, 11 for Signal}
}
\value{
Submission status and stats
}
\description{
Submission status of the last submission associated with the account
}
\examples{
\dontrun{
model_id = get_models()[["bayo"]]
submission_status(model_id = model_id,tournament=8)
}
}
