% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{submit_predictions_multi}
\alias{submit_predictions_multi}
\title{Function to submit the Numerai Tournament predictions for multiple tournaments}
\usage{
submit_predictions_multi(submissions, location = tempdir())
}
\arguments{
\item{submissions}{The named list of the data frames of predictions to submit. The list names should be tournament names and each element a data frame having two columns named "id" & "probability" for that particular tournament}

\item{location}{The location in which to store the predictions}
}
\value{
The submission ids for the submissions made
}
\description{
Function to submit the Numerai Tournament predictions for multiple tournaments
}
\examples{
\dontrun{
submission_ids <- submit_predictions_multi(submissions_data)
}
}
