% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signalsapi.R
\name{ticker_universe}
\alias{ticker_universe}
\title{Fetch universe of accepted tickers}
\usage{
ticker_universe()
}
\value{
List of currently accepted tickers
}
\description{
Fetch universe of accepted tickers
}
\examples{
\dontrun{
ticker_universe()
}
}
