% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseapi.R
\name{upload_diagnostics}
\alias{upload_diagnostics}
\title{Upload predictions to diagnostics from file.}
\usage{
upload_diagnostics(file_path = NA, model_id, df = NA, tournament = 8)
}
\arguments{
\item{file_path}{CSV file with predictions that will get uploaded}

\item{model_id}{Target model UUID (required for accounts with multiple models)}

\item{df}{DataFrame to upload, if given both df and file_path, df will be uploaded.}

\item{tournament}{Tournament ID, 8 for Main, 11 for Signal}
}
\value{
diagnostics_id
}
\description{
Upload predictions to diagnostics from file.
}
\examples{
\dontrun{
upload_diagnostics(file_path = "prediction.csv",model_id = get_models()[["bayo"]])
}
}
