% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{user_info}
\alias{user_info}
\title{Get information about your username}
\usage{
user_info(model_id = NULL)
}
\arguments{
\item{model_id}{The id of the model}
}
\value{
A list containing information about user
}
\description{
Get information about your username
}
\examples{
\dontrun{
uinfo <- user_info()
names(uinfo)
uinfo$Latest_Submission
}
}
