% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{user_performance}
\alias{user_performance}
\title{Get User Performance}
\usage{
user_performance(user_name = "theomniacs")
}
\arguments{
\item{user_name}{UserName for which performance metrics to get}
}
\value{
Get User Performance
}
\description{
Get User Performance
}
\examples{
\dontrun{
user_performance(user_name="theomniacs")
}
}
