% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseapi.R
\name{wallet_transactions}
\alias{wallet_transactions}
\title{Get all transactions in your wallet.}
\usage{
wallet_transactions()
}
\value{
Wallet Txs Data Frame
}
\description{
Get all transactions in your wallet.
}
\examples{
\dontrun{
wallet_transactions()
}
}
