% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{download_data}
\alias{download_data}
\title{Function to download the Numerai Tournament data}
\usage{
download_data(location = tempdir(), tournament = "NOMI")
}
\arguments{
\item{location}{The directory path in which to store the data}

\item{tournament}{The name of the tournament, Default is Nomi and is not case-sensitive. Since at the moment the datasets are same for all tournaments this parameter can be left blank.}
}
\value{
A list containing the training and tournament data objects
}
\description{
Function to download the Numerai Tournament data
}
\examples{
\dontrun{
## Directory where data files and prediction files to be saved
## Put custom directory path or use the current working directory
data_dir <- tempdir()

## Download data set for current competition
data <- download_data(data_dir,tournament="Nomi")
data_train <- data$data_train
data_tournament <- data$data_tournament
}
}
