% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{get_models}
\alias{get_models}
\title{Get models associated with your account}
\usage{
get_models()
}
\value{
A list containing information about the models
}
\description{
Get models associated with your account
}
\examples{
\dontrun{
models <- get_models()
}
}
