% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{performance_distribution}
\alias{performance_distribution}
\title{Get the performance of the user as a distribution}
\usage{
performance_distribution(username, metric)
}
\description{
Get the performance of the user as a distribution
}
