% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{performance_distribution}
\alias{performance_distribution}
\title{Get the performance of the user as a distribution}
\usage{
performance_distribution(
  username,
  metric,
  merge = FALSE,
  round_aggregate = TRUE
)
}
\arguments{
\item{username}{A vector of one or more usernames}

\item{metric}{A statistic, as a character vector.}

\item{merge}{If TRUE, combine the usernames into a single result}

\item{round_aggregate}{If TRUE, aggregate the submission data by round}
}
\description{
Get the performance of the user as a distribution
}
