% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{performance_over_time}
\alias{performance_over_time}
\alias{user_performance_data}
\title{Get the performance of the user over time}
\usage{
user_performance_data(username)

performance_over_time(username, metric)
}
\description{
Get the performance of the user over time

Get the performance of the user over time
}
