% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{round_stats}
\alias{round_stats}
\title{Get Information for a Round Number}
\usage{
round_stats(round_number, tournament = "Nomi")
}
\arguments{
\item{round_number}{Round Number for which information to fetch}

\item{tournament}{The name of the tournament, Default is Nomi and is not case-sensitive}
}
\value{
List containing general round information
}
\description{
Get Information for a Round Number
}
\examples{
\dontrun{
round_stats(round_number=177)
}
}
