% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseapi.R
\name{run_query}
\alias{run_query}
\title{Run a custom query}
\usage{
run_query(query, auth = TRUE)
}
\arguments{
\item{query}{Query to pass as a string}

\item{auth}{Whether the query need authorisation to run}
}
\value{
Parsed result from custom query
}
\description{
Run a custom query
}
\examples{
\dontrun{
run_query(query = 'query{account{username }}', auth=TRUE)
}
}
