% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{stake_nmr}
\alias{stake_nmr}
\title{Stake NMR on the current round and single tournament}
\usage{
stake_nmr(tournament = "Bernie", value, confidence, mfa_code = "",
  password = "")
}
\arguments{
\item{tournament}{The name of the tournament, Default is Bernie and is not case-sensitive}

\item{value}{The amount of NMR to stake}

\item{confidence}{The confidence value to use}

\item{mfa_code}{The mfa code}

\item{password}{Your password}
}
\value{
The transaction hash for stake made
}
\description{
Stake NMR on the current round and single tournament
}
\examples{
\dontrun{
stake_tx_hash <- stake_nmr(tournament="Elizabeth",value = 1, confidence = ".5")
}
}
