% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{submit_predictions}
\alias{submit_predictions}
\title{Function to submit the Numerai Tournament predictions}
\usage{
submit_predictions(
  submission,
  model_id = NULL,
  location = tempdir(),
  tournament = "Kazutsugi"
)
}
\arguments{
\item{submission}{The data frame of predictions to submit. This should have two columns named "id" & "prediction_kazutsugi"}

\item{model_id}{Target model UUID (required for accounts with multiple models)}

\item{location}{The location in which to store the predictions}

\item{tournament}{The name of the tournament, Default is Kazutsugi and is not case-sensitive}
}
\value{
The submission id for the submission made
}
\description{
Function to submit the Numerai Tournament predictions
}
\examples{
\dontrun{
submission_id <- submit_predictions(submission_data,tournament="Kazutsugi")
}
}
