% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{submit_predictions}
\alias{submit_predictions}
\title{Function to submit the Numerai Tournament predictions}
\usage{
submit_predictions(submission, location = tempdir())
}
\arguments{
\item{submission}{The data frame of predictions to submit. This should have two columns named "id" & "probability"}

\item{location}{The location in which to store the predictions}
}
\value{
The submission id for the submission made
}
\description{
Function to submit the Numerai Tournament predictions
}
\examples{
\dontrun{
submission_id <- submit_predictions(submission_data)
}
}
